requirements.txt
anyio==4.5.2
AnyQt==0.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.15.8
asttokens==2.0.5
attrs==23.2.0
backcall==0.2.0
backports.tarfile==1.2.0
baycomp==1.0.3
beautifulsoup4==4.11.1
biopython==1.83
bioservices==1.11.2
bleach==5.0.1
Bottleneck==1.4.0
catboost==1.2.7
cattrs==23.2.3
certifi==2024.2.2
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
commonmark==0.9.1
coverage==7.6.1
cycler==0.11.0
datacache==1.4.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.4.0
docopt==0.6.2
docopt-ng==0.9.0
docutils==0.20.1
easydev==0.13.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
execnet==2.1.2
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
ftpretty==0.4.0
future==1.0.0
Genesis-PyAPI==1.2.1
gevent==24.2.1
goatools==1.5.2
graphql-core==3.2.6
graphviz==0.20.3
greenlet==3.0.3
grequests==0.7.0
gtfparse==2.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.6
igraph==0.11.4
importlib-resources==5.8.0
importlib_metadata==7.0.2
iniconfig==2.1.0
intermine==1.13.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.13.2
jaraco.classes==3.3.1
jaraco.context==6.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keyring==24.3.1
keyrings.alt==5.0.2
kiwisolver==1.4.3
lazy-object-proxy==1.10.0
line-profiler==4.1.2
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
matplotlib-venn==1.1.2
mccabe==0.7.0
mdurl==0.1.2
memoized-property==1.0.3
mistune==0.8.4
mock==5.2.0
more-itertools==10.2.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nbval==0.11.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
numpy==1.23.0
openpyxl==3.1.2
openTSNE==1.0.1
Orange-Bioinformatics==2.6.25
orange-canvas-core==0.2.2
orange-widget-base==4.24.0
Orange3==3.37.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==4.2.0
plotly==5.24.1
pluggy==1.5.0
polars==0.20.31
progressbar33==2.4
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.21
pydot==3.0.2
pyensembl==2.3.13
Pygments==2.18.0
pylint==2.17.7
PyMySQL==1.1.0
pyparsing==3.0.9
pypdb==2.3
pyqtgraph==0.13.3
pyrsistent==0.18.1
pytest==8.3.5
pytest-xdist==3.6.1
python-dateutil==2.8.2
python-igraph==0.11.4
python-louvain==0.16
pytz==2022.1
PyYAML==6.0.2
pyzmq==23.2.0
qasync==0.27.1
qtconsole==5.3.1
QtPy==2.1.0
rcsb-api==1.1.3
requests==2.31.0
requests-cache==1.2.0
rich==13.9.3
rustworkx==0.15.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.11.2
Send2Trash==1.8.0
serializable==0.4.1
serverfiles==0.3.1
simplejson==3.20.2
six==1.16.0
slumber==0.7.1
sniffio==1.3.1
soupsieve==2.3.2.post1
stack-data==0.3.0
statsmodels==0.14.1
suds-community==1.1.2
tenacity==9.0.0
terminado==0.15.0
texttable==1.7.0
threadpoolctl==3.5.0
tinycss2==1.1.1
tinytimer==0.0.0
tomli==2.3.0
tomlkit==0.13.3
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typechecks==0.1.0
typing_extensions==4.10.0
UpSetPlot==0.9.0
url-normalize==1.4.3
urllib3==2.2.1
venn==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
wrapt==1.16.0
xgboost==2.1.2
xlrd==2.0.1
XlsxWriter==3.2.0
xmltodict==0.13.0
zipp==3.8.0
zope.event==5.0
zope.interface==6.2
